Rename SimplifyQuoteEscapeRector
