+++ rules-tests/CodingStyle/Rector/String_/SimplifyQuoteEscapeRector/Fixture/skip.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\CodingStyle\Rector\String_\SymplifyQuoteEscapeRector\Fixture;
+namespace Rector\Tests\CodingStyle\Rector\String_\SimplifyQuoteEscapeRector\Fixture;
 
 class Skip
 {
