+++ rules-tests/CodingStyle/Rector/String_/SimplifyQuoteEscapeRector/config/configured_rule.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+use Rector\CodingStyle\Rector\String_\SimplifyQuoteEscapeRector;
+use Rector\Config\RectorConfig;
+
+return RectorConfig::configure()
+    ->withRules([SimplifyQuoteEscapeRector::class]);
