+++ rules/CodingStyle/Rector/String_/SimplifyQuoteEscapeRector.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\CodingStyle\Rector\String_;
+
+/**
+ * @see \Rector\Tests\CodingStyle\Rector\String_\SimplifyQuoteEscapeRector\SimplifyQuoteEscapeRectorTest
+ *
+ * @phpstan-ignore class.extendsDeprecatedClass,symplify.forbiddenExtendOfNonAbstractClass
+ */
+final class SimplifyQuoteEscapeRector extends SymplifyQuoteEscapeRector
+{
+}
